+++ src/Implementation/Format/ExcelFormat.php
@@ -7,8 +7,8 @@ use ILIAS\UI\Renderer;
 use srag\DataTable\Component\Column\Column;
 use srag\DataTable\Component\Data\Data;
 use srag\DataTable\Component\Data\Row\RowData;
-use srag\DataTable\Component\Table;
 use srag\DataTable\Component\Settings\Settings;
+use srag\DataTable\Component\Table;
 
 /**
  * Class ExcelFormat
@@ -103,11 +103,11 @@ class ExcelFormat extends AbstractFormat
     /**
      * @inheritDoc
      */
-    protected function handleRow(Table $component, array $columns, RowData $row, Settings $settings, Renderer $renderer) : void
+    protected function handleRow(Table $component, array $columns, RowData $row, Renderer $renderer) : void
     {
         $this->current_col = 0;
 
-        parent::handleRow($component, $columns, $row, $settings, $renderer);
+        parent::handleRow($component, $columns, $row, $renderer);
 
         $this->current_row++;
     }